raise base deal limit

+++ binancespotbot/src/main/java/ru/tyumentsev/binancespotbot/cache/MarketData.java
@@ -60,7 +60,7 @@ public class MarketData {
 
     @NonFinal
     @Value("${strategy.global.maximalPairPrice}")
-    int maximalPairPrice;
+    double maximalPairPrice;
     @NonFinal
     @Value("${strategy.global.candlestickEventsCacheSize}")
     int candlestickEventsCacheSize;
